Add constructors to memory backing
